# Support for custom character length alphabets