--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -9,7 +9,7 @@ jobs:
     strategy:
       matrix:
         platform: [ubuntu-latest, macos-latest, windows-latest]
-        go-version: ['1.16', '1.17']
+        go-version: ['1.18']
     runs-on: ${{ matrix.platform }}
     steps:
       - name: Setup Go