--- a/alphabet_test.go
+++ b/alphabet_test.go
@@ -1,22 +1,23 @@
 package shortuuid
 
 import (
-	"strings"
+	"sort"
 	"testing"
 )
 
 func TestDedupe(t *testing.T) {
 	tests := []struct {
-		in, out string
+		in, out []rune
 	}{
-		{"01010101010101", "01"},
-		{"abcabcfoo", "abcfo"},
+		{[]rune("01010101010101"), []rune("01")},
+		{[]rune("abcabcfoo"), []rune("abcfo")},
 	}
 
 	for _, test := range tests {
-		in := strings.Join(dedupe(strings.Split(test.in, "")), "")
-		if in != test.out {
-			t.Errorf("expected %q, got %q", in, test.out)
+		sort.Slice(test.in, func(i, j int) bool { return test.in[i] < test.in[j] })
+		in := dedupe(test.in)
+		if string(in) != string(test.out) {
+			t.Errorf("expected %q, got %q", string(test.out), string(in))
 		}
 	}
 }
@@ -48,5 +49,14 @@ func TestAlphabetIndexError(t *testing.T) {
 	idx, err := abc.Index('l')
 	if err == nil {
 		t.Errorf("expected an error, got a valid index %d", idx)
+	}
+}
+
+func BenchmarkAlphabetIndex(b *testing.B) {
+	abc := newAlphabet(DefaultAlphabet)
+	for i := 0; i < b.N; i++ {
+		for _, ch := range DefaultAlphabet {
+			_, _ = abc.Index(ch)
+		}
 	}
 }