--- a/alphabet_test.go
+++ b/alphabet_test.go
@@ -4,23 +4,6 @@ import (
 	"testing"
 )
 
-func TestDedupe(t *testing.T) {
-	tests := []struct {
-		in, out []rune
-	}{
-		{[]rune("01010101010101"), []rune("01")},
-		{[]rune("abcabcfoo"), []rune("abcfo")},
-	}
-
-	for _, test := range tests {
-		sortRunes(test.in)
-		in := dedupe(test.in)
-		if string(in) != string(test.out) {
-			t.Errorf("expected %q, got %q", string(test.out), string(in))
-		}
-	}
-}
-
 func TestAlphabetIndex(t *testing.T) {
 	abc := newAlphabet(DefaultAlphabet)
 	idx, err := abc.Index('z')