--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,6 @@ module github.com/lithammer/shortuuid/v4
 
 require github.com/google/uuid v1.6.0
 
-go 1.13
+require golang.org/x/exp v0.0.0-20220827204233-334a2380cb91
+
+go 1.18