--- a/shortuuid.go
+++ b/shortuuid.go
@@ -33,9 +33,9 @@ func NewWithNamespace(name string) string {
 	switch {
 	case name == "":
 		u = uuid.New()
-	case strings.HasPrefix(strings.ToLower(name), "http://"):
+	case hasPrefixCaseInsensitive(name, "https://"):
 		u = uuid.NewSHA1(uuid.NameSpaceURL, []byte(name))
-	case strings.HasPrefix(strings.ToLower(name), "https://"):
+	case hasPrefixCaseInsensitive(name, "http://"):
 		u = uuid.NewSHA1(uuid.NameSpaceURL, []byte(name))
 	default:
 		u = uuid.NewSHA1(uuid.NameSpaceDNS, []byte(name))
@@ -49,4 +49,8 @@ func NewWithNamespace(name string) string {
 func NewWithAlphabet(abc string) string {
 	enc := encoder{newAlphabet(abc)}
 	return enc.Encode(uuid.New())
+}
+
+func hasPrefixCaseInsensitive(s, prefix string) bool {
+	return len(s) >= len(prefix) && strings.EqualFold(s[:len(prefix)], prefix)
 }