- resolve #39
(compatible with python's shortuuid custom len alphabets)

- resolve #60

- Improved performance:
```
goos: linux
goarch: amd64
cpu: 12th Gen Intel(R) Core(TM) i5-12400F
                         │    v4.1.0    │                new                  │
                         │    sec/op    │   sec/op     vs base                │
UUID-12                     412.4n ± 0%   326.5n ± 0%  -20.84% (p=0.000 n=10)
Encoding-12                136.00n ± 1%   43.73n ± 0%  -67.85% (p=0.000 n=10)
Decoding-12                 270.3n ± 0%   180.0n ± 1%  -33.42% (p=0.000 n=10)
NewWithAlphabet-12         8064.0n ± 0%   623.8n ± 0%  -92.27% (p=0.000 n=10)
EncodingB57_MB-12                         122.2n ± 5%
EncodingB16-12                            132.5n ± 1%
EncodingB16_MB-12                         358.6n ± 0%
DecodingB16-12                            220.4n ± 1%
DecodingB16_MB-12                         266.8n ± 2%
NewWithAlphabetB16-12                     486.5n ± 0%
NewWithAlphabetB16_MB-12                  837.2n ± 0%

                         │     B/op      │    B/op     vs base                  │
UUID-12                     40.00 ± 0%     40.00 ± 0%        ~ (p=1.000 n=10) ¹
Encoding-12                 24.00 ± 0%     24.00 ± 0%        ~ (p=1.000 n=10) ¹
Decoding-12                 0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
NewWithAlphabet-12         8545.5 ± 0%     280.0 ± 0%  -96.72% (p=0.000 n=10)
EncodingB57_MB-12                          72.00 ± 0%
EncodingB16-12                             64.00 ± 0%
EncodingB16_MB-12                          512.0 ± 0%
DecodingB16-12                             0.000 ± 0%
DecodingB16_MB-12                          0.000 ± 0%
NewWithAlphabetB16-12                      144.0 ± 0%
NewWithAlphabetB16_MB-12                   592.0 ± 0%

                         │   allocs/op   │ allocs/op   vs base                  │
UUID-12                     2.000 ± 0%     2.000 ± 0%        ~ (p=1.000 n=10) ¹
Encoding-12                 1.000 ± 0%     1.000 ± 0%        ~ (p=1.000 n=10) ¹
Decoding-12                 0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
NewWithAlphabet-12         19.000 ± 0%     3.000 ± 0%  -84.21% (p=0.000 n=10)
EncodingB57_MB-12                          2.000 ± 0%
EncodingB16-12                             2.000 ± 0%
EncodingB16_MB-12                          4.000 ± 0%
DecodingB16-12                             0.000 ± 0%
DecodingB16_MB-12                          0.000 ± 0%
NewWithAlphabetB16-12                      4.000 ± 0%
NewWithAlphabetB16_MB-12                   6.000 ± 0%
```