@anatoly-kussul 2024-11-30
Changed `alphabet.Index()` to use binary search, which gains ~10% performance on Decode:
```
                   │     for     │                binary               │
                   │   sec/op    │   sec/op     vs base                │
Decoding-12          203.4n ± 1%   180.2n ± 1%  -11.45% (p=0.000 n=10)
```
Previous attempt in #51 didn't find any success, because it used `sort.Search` which adds lambda calls overhead.
```

                   │     for     │             sort.Search             │
                   │   sec/op    │   sec/op     vs base                │
Decoding-12          203.4n ± 1%   313.1n ± 1%  +53.92% (p=0.000 n=10)
```