@anatoly-kussul 2024-12-01
Added some more benchmarks,
Added optimization for alphabets containing only single-byte characters.