@lithammer 2024-12-01
> Had to copy some internal Sort's code into project, as before go1.18 `sort.Slice` uses reflect, which increases allocations and decreases performance

I think we can just raise the minimum version in `go.mod` to 1.18 instead. 1.13 has been EOL for over 4 years. And 1.18 for almost 2 years. 

https://github.com/lithammer/shortuuid/blob/d3d4af740f99283d2e1759a5fa2b96cdf1f9927f/go.mod#L5