@anatoly-kussul 2024-12-01
seems like I was a bit mistaken, and `slices` package was an external experimental package at go 1.18, so will have to introduce it as external dependency.
Or if it's ok, we can increase minimum version up to 1.21, slices are already built-in at that point. (1.21 EoL was 3 months ago).

And no clue why CI lint is failing, have to investigate