@anatoly-kussul 2024-12-01
Introduced more encode optimizations, using similar technique as used in `big.Int.Text()`.
Updated benchmark in PR description.