@anatoly-kussul 2024-12-02
Added benchmarks for NewWithNamespace.
Increase preformance of it by removing the need to lower full string before comparison.
```
goos: linux
goarch: amd64
cpu: 12th Gen Intel(R) Core(TM) i5-12400F
                         │     old      │                 new                 │
                         │    sec/op    │   sec/op     vs base                │
NewWithNamespace-12        285.4n ±  1%   248.7n ± 2%  -12.89% (p=0.000 n=10)
NewWithNamespaceHttp-12    281.4n ±  1%   253.6n ± 1%   -9.88% (p=0.000 n=10)
NewWithNamespaceHttps-12   322.8n ± 20%   253.6n ± 2%  -21.41% (p=0.000 n=10)
geomean                    296.0n         252.0n       -14.87%

                         │    B/op    │    B/op     vs base                 │
NewWithNamespace-12        200.0 ± 0%   200.0 ± 0%       ~ (p=1.000 n=10)
NewWithNamespaceHttp-12    208.0 ± 0%   208.0 ± 0%       ~ (p=1.000 n=10)
NewWithNamespaceHttps-12   224.0 ± 0%   224.0 ± 0%       ~ (p=1.000 n=10)
geomean                    210.4        210.4       +0.00%

                         │ allocs/op  │ allocs/op   vs base                 │
NewWithNamespace-12        5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=10)
NewWithNamespaceHttp-12    5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=10)
NewWithNamespaceHttps-12   5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=10)
geomean                    5.000        5.000       +0.00%
```