@anatoly-kussul 2024-12-02
> Thanks! Very impressive work again! 👏

Thanks again for very quick review and approval!

> Are you planning on doing more of these or...?

I think I did all I could already. =)

I just noticed that this library (we used v3) was major part of our cpu usage on production.
So I tried to improve performance of it a bit.
We only use `New` and `NewWithNamespace`, so that was my main focus in first PR.
But then I noticed, that I also can improve `NewWithAlphabet` by a lot, and decided to give it a try.