@lithammer 2024-12-02
Well I'm glad you took your time. The improvements are quite significant! Always nice to hear from people using it in production 🙂 

New release: https://github.com/lithammer/shortuuid/releases/tag/v4.2.0